Converting data from Tensorflow to NumPy Array
This code is useful if you are training the model on your hosted machine without the relevant GPU support. This is optional and feel free to skip it if you have the support for the GPU but keep in view that if you are not converting the code to the NumPy array and using the Tensor make sure that it is consistent with the rest of the code to avert problems in the execution of the notebook.